feat: steering UX wave — calm landing + drawer, one Add menu; retire policies/domain/coverage settings panels
Direct user feedback on the Steering surface: "It's a lot to look at."
Presentation only — every wire (src/api/steering.ts, the shipping rules
CRUD, the wiki scoreboard/meta reads) is untouched.

D1 — the Steering surface redesigned:
- /steering is a LANDING now: a calm grid of seven compact type cards,
  each counting that type's rules from the one rules fetch (client-side,
  the same steeringTypeOf fold the pages use); card click -> /steering/<type>.
- /steering/<type> is a clean rule LIST — severity chip, id + one-line
  statement, weight only when non-default. Nothing renders open by default.
- ONE "Add" menu per type page holds the three management flows — Add
  individual (modal form), Import (file-picker panel), Add with chat (the
  governed authoring run) — each opening on demand. The menu items keep the
  retired management bar's testids (the affordances survived).
- A row click opens the rule DRAWER: full statement, provenance,
  applies_to/excludes/weight/effect, evidence, retire/edit — everything
  that used to crowd the page inline. The retire kill switch (typed
  confirmation + required reason) is unchanged, just opened from here.
- SteeringPage.tsx decomposed from a 1565-line monolith into a 291-line
  shell plus SteeringTypeCards, SteeringRuleList, SteeringRuleDrawer,
  SteeringAddMenu, SteeringRuleForm, SteeringImportPanel,
  SteeringAuthorPanel, SteeringHealth, SteeringChips, fileText.
  AuthorPanel/readFileText moved (TestingPage re-imports; reused verbatim).
- The AW-23 health header keeps its per-type/store-wide/501 honesty states.

D2 — Policies removed from Settings: the 'policies' panel and
PolicyManager.tsx (+ its test) deleted — policies merged into steering
rules, the Steering pages are the surface. /policies redirects (replace)
to /steering; the rail maps it to the Steering heading on the
pre-redirect tick; the palette's claim fallback now targets /steering.

D3 — the orphaned 'domain' and 'coverage' Settings panels removed:
DomainModelBrowser.tsx deleted (nothing else used it); CoverageView.tsx
STAYS (RightPanel renders it per-run). /domain and /coverage redirect
(replace) to /system. Settings rail rows are Theme / Workflows / System.

Routes: bare /steering never redirects any more (it IS the landing);
/wiki, /rules, /policies and typo'd /steering/* replace onto /steering;
'policies'/'coverage'/'domain' left the Panel union.

Tests: landing cards + counts + nav, breadcrumb, Add-menu opens each
flow, drawer detail + honest defaults, retire from the drawer, redirects,
Settings absences, rail order Testing -> Steering -> Settings unchanged.
vitest 1975 passed / tsc clean / eslint clean / vite build ok;
testid inventory regenerated with the repo script.
