+++ tests/SettingsRailSection.test.tsx
@@ -15,18 +15,27 @@ import { version as pkgVersion } from '../package.json';
 afterEach(cleanup);
 
 describe('SettingsShortcutRows', () => {
-  it('carries the retired AppChrome dropdown entries MINUS the two Steering absorbed', () => {
+  it('carries the retired AppChrome dropdown entries MINUS everything governance retired elsewhere', () => {
     // The §1.2 contract: the rows hold the dropdown's entries, so both the
     // /system Settings entry AND the /theme entry (PR #64) must be present.
     expect(SETTINGS_ITEMS).toContainEqual({ label: 'System', path: '/system' });
     expect(SETTINGS_ITEMS).toContainEqual({ label: 'Theme', path: '/theme' });
+    expect(SETTINGS_ITEMS).toContainEqual({ label: 'Workflows', path: '/workflows' });
     // The STEERING program: Rules (/rules) and Arch Wiki (/wiki) retired into
-    // the Steering primary path — their rows are GONE from Settings, and the
-    // old addresses redirect to /steering/architecture.
+    // the Steering primary path — their rows are GONE from Settings.
     expect(SETTINGS_ITEMS.map((i) => i.label)).not.toContain('Rules');
     expect(SETTINGS_ITEMS.map((i) => i.label)).not.toContain('Arch Wiki');
     expect(SETTINGS_ITEMS.map((i) => i.path)).not.toContain('/rules');
     expect(SETTINGS_ITEMS.map((i) => i.path)).not.toContain('/wiki');
+    // The steering-UX wave: Policies merged into steering rules (the address
+    // redirects to /steering), and the orphaned Coverage/Domain panels retired
+    // (both redirect to /system) — none of the three is a Settings row any more.
+    expect(SETTINGS_ITEMS.map((i) => i.label)).not.toContain('Policies');
+    expect(SETTINGS_ITEMS.map((i) => i.label)).not.toContain('Coverage');
+    expect(SETTINGS_ITEMS.map((i) => i.label)).not.toContain('Domain');
+    expect(SETTINGS_ITEMS.map((i) => i.path)).not.toContain('/policies');
+    expect(SETTINGS_ITEMS.map((i) => i.path)).not.toContain('/coverage');
+    expect(SETTINGS_ITEMS.map((i) => i.path)).not.toContain('/domain');
 
     render(<SettingsShortcutRows navigate={() => {}} />);
     expect(screen.getByRole('menu')).toBeInTheDocument();
